+++ webcomponents/slides/poll/src/components/slide/deckdeckgo-slide-poll.tsx
@@ -14,7 +14,7 @@ import {CommunicationService} from '../../services/communication/communication.s
 @Component({
   tag: 'deckgo-slide-poll',
   styleUrl: 'deckdeckgo-slide-poll.scss',
-  shadow: true
+  shadow: true,
 })
 export class DeckdeckgoSlidePoll implements DeckdeckgoSlideResize {
   @Element() el: HTMLElement;
@@ -33,9 +33,6 @@ export class DeckdeckgoSlidePoll implements DeckdeckgoSlideResize {
 
   @Prop({reflect: true, mutable: true})
   pollKey: string;
-
-  @Prop({reflectToAttr: true}) customActions: boolean = false;
-  @Prop({reflectToAttr: true}) customBackground: boolean = false;
 
   private answerSlots: string[];
 
@@ -151,8 +148,8 @@ export class DeckdeckgoSlidePoll implements DeckdeckgoSlideResize {
       this.chartData = [
         {
           label: question ? question.innerHTML : 'Poll',
-          values: activeBars
-        }
+          values: activeBars,
+        },
       ];
 
       this.chartData = [...this.chartData];
@@ -228,7 +225,7 @@ export class DeckdeckgoSlidePoll implements DeckdeckgoSlideResize {
         {
           label: this.chartData[0].label as string,
           values: this.chartData[0].values,
-          answered: this.answeredOnce
+          answered: this.answeredOnce,
         },
         this.updatePollKeyCallback,
         this.updateVoteCallback
@@ -256,7 +253,7 @@ export class DeckdeckgoSlidePoll implements DeckdeckgoSlideResize {
         {
           label: this.chartData[0].label as string,
           values: this.chartData[0].values,
-          answered: this.answeredOnce
+          answered: this.answeredOnce,
         },
         this.pollKey
       );
@@ -296,7 +293,7 @@ export class DeckdeckgoSlidePoll implements DeckdeckgoSlideResize {
       this.chartData = [];
       this.chartData.push({
         label: poll.poll.label,
-        values: poll.poll.values as DeckdeckgoBarChartDataValue[]
+        values: poll.poll.values as DeckdeckgoBarChartDataValue[],
       });
 
       await this.initAnswersData();
@@ -427,6 +424,8 @@ export class DeckdeckgoSlidePoll implements DeckdeckgoSlideResize {
           <slot name="notes"></slot>
           <slot name="actions"></slot>
           <slot name="background"></slot>
+          <slot name="header"></slot>
+          <slot name="footer"></slot>
         </div>
       </Host>
     );